% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{app_server}
\alias{app_server}
\title{Application server}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input, output, session}{Shiny server arguments}
}
\description{
Server for the production \code{dfdiffs} comparison app
}
