% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react_themes.R
\name{base_react_theme}
\alias{base_react_theme}
\title{reactable theme (base data)}
\usage{
base_react_theme(dark = FALSE)
}
\arguments{
\item{dark}{use dark-mode colors (default FALSE)}
}
\description{
theme for base data in reactable table
}
