% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmrn_fresh_theme.R
\name{bmrn_fresh_theme}
\alias{bmrn_fresh_theme}
\title{BioMarin theme (fresh <> bs4Dash)}
\usage{
bmrn_fresh_theme()
}
\value{
theme shiny app
}
\description{
this is the fresh theme with BioMarin colors.
}
