% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react_themes.R
\name{changed_react_theme}
\alias{changed_react_theme}
\title{reactable theme (changed data)}
\usage{
changed_react_theme(dark = FALSE)
}
\arguments{
\item{dark}{use dark-mode colors (default FALSE)}
}
\description{
theme for changed data in reactable table
}
