% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react_themes.R
\name{comp_react_theme}
\alias{comp_react_theme}
\title{reactable theme (compare data)}
\usage{
comp_react_theme(dark = FALSE)
}
\arguments{
\item{dark}{use dark-mode colors (default FALSE)}
}
\description{
theme for compare data in reactable table
}
