% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_deleted_data.R
\name{create_deleted_data}
\alias{create_deleted_data}
\title{Create a dataset of deleted records}
\usage{
create_deleted_data(compare, base, by = NULL, by_col = NULL, cols = NULL)
}
\arguments{
\item{compare}{a 'new' or 'current' dataset}

\item{base}{an 'old' or 'previous' dataset}

\item{by}{the joining bs4Dash::column between the two datasets}

\item{by_col}{name of the new joining bs4Dash::column}

\item{cols}{names of columns to compare}
}
\value{
deleted_data
}
\description{
Create a dataset of deleted records
}
\examples{
# using local data
CompleteData <- dfdiffs::CompleteData
IncompleteData <- dfdiffs::IncompleteData
create_deleted_data(compare = IncompleteData,
                    base = CompleteData)
}
