% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_empty_tbl.R
\name{create_empty_tbl}
\alias{create_empty_tbl}
\title{Create empty tibble}
\usage{
create_empty_tbl(tbl)
}
\arguments{
\item{tbl}{input table}
}
\value{
tibble with columns from tbl, all logical
}
\description{
Create empty tibble
}
\examples{
DF <- data.frame(
        id = 1:6,
        lowercase = letters[1:6],
        uppercase = LETTERS[1:6])
create_empty_tbl(DF)
}
