% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_join_column.R
\name{create_join_column}
\alias{create_join_column}
\title{Create unique row identifier}
\usage{
create_join_column(df, by_colums, new_by_column_name)
}
\arguments{
\item{df}{a data.frame or tibble}

\item{by_colums}{columns to uniquely identify a row}

\item{new_by_column_name}{the new column name}
}
\value{
join_col_data
}
\description{
Create unique row identifier
}
\examples{
# using dfdiffs::diff_modified_data
diff_modified_data <- dfdiffs::diff_modified_data
current_modified <- diff_modified_data$diff_current_modified
previous_modified <- diff_modified_data$diff_previous_modified
create_join_column(df = current_modified,
                   by_columns = c("subject_id", "record"),
                   new_by_column_name = "join")
}
