% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_data.R
\name{create_new_data}
\alias{create_new_data}
\title{Create a table of new records}
\usage{
create_new_data(compare, base, by = NULL, by_col = NULL, cols = NULL)
}
\arguments{
\item{compare}{a 'new' or 'current' dataset}

\item{base}{an 'old' or 'previous' dataset}

\item{by}{the joining bs4Dash::column between the two datasets}

\item{by_col}{name of the new joining bs4Dash::column}

\item{cols}{names of columns to compare}
}
\value{
new_data
}
\description{
Create a table of new records
}
\examples{
# using local data
T2Data <- dfdiffs::T2Data
T1Data <- dfdiffs::T1Data
create_new_data(compare = T2Data, base = T1Data,
                by = c('subject', 'record'))
create_new_data(compare = T2Data, base = T1Data,
                by = c('subject', 'record'),
                cols = c("text_var", "factor_var"))
}
