% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_tabyl.R
\name{cross_tabyl}
\alias{cross_tabyl}
\title{cross_tabyl}
\usage{
cross_tabyl(df, col)
}
\arguments{
\item{df}{a data.frame or tibble}

\item{col}{a column with categorical or factor data}
}
\value{
cross_tabyl
}
\description{
cross_tabyl() is a helper function to quickly create a
cross-tabulation of the values in a categorical variable.
}
\examples{
# not run
df <- data.frame(hair_color = c("blond", "brown", "brown", "black", NA))
cross_tabyl(df, "hair_color")
}
