% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react_themes.R
\name{deleted_react_theme}
\alias{deleted_react_theme}
\title{reactable theme (deleted data)}
\usage{
deleted_react_theme(dark = FALSE)
}
\arguments{
\item{dark}{use dark-mode colors (default FALSE)}
}
\description{
theme for deleted data in reactable table
}
