% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_compareDataServer.R
\name{dev_compareDataServer}
\alias{dev_compareDataServer}
\title{dev_compareDataServer}
\usage{
dev_compareDataServer(id, data_selected)
}
\arguments{
\item{id}{module id}
}
\description{
dev compare server module
}
