% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_selectDataServer.R
\name{dev_selectDataServer}
\alias{dev_selectDataServer}
\title{dev_selectDataServer}
\usage{
dev_selectDataServer(id, data_upload)
}
\arguments{
\item{id}{module id}
}
\description{
dev select server module
}
