% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_uploadDataServer.R
\name{dev_uploadDataServer}
\alias{dev_uploadDataServer}
\title{dev_uploadDataServer()}
\usage{
dev_uploadDataServer(id)
}
\arguments{
\item{id}{module id}
}
\description{
development Server module for upload
}
