% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayDataServer.R
\name{displayDataServer}
\alias{displayDataServer}
\title{Display uploaded data (Server)}
\usage{
displayDataServer(id, data_upload)
}
\arguments{
\item{id}{module id}

\item{data_upload}{list or uploaded xlsx and flat files}
}
\value{
list of previous xlsx/flat files and current xlsx/flat files
}
\description{
Display uploaded data (Server)
}
