% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_compare_report.R
\name{generate_compare_report}
\alias{generate_compare_report}
\title{Generate comparison report (xlsx)}
\usage{
generate_compare_report(
  compare,
  base,
  by = NULL,
  by_col = NULL,
  cols = NULL,
  path
)
}
\arguments{
\item{compare}{comparison data table}

\item{base}{base data table}

\item{by}{join column}

\item{by_col}{new join column name}

\item{cols}{columns to compare}

\item{path}{path to excel file}
}
\value{
exported xlsx file
}
\description{
Generate comparison report (xlsx)
}
\examples{
# not run
m15 <- dfdiffs::master15
m22 <- dfdiffs::master22
generate_compare_report(compare = m22, base = m15,
    by = "playerID",
    by_col = "join",
    cols = c("nameFirst", "nameLast", "nameGiven"),
    path = "inst/extdata/xlsx/compare-report-text.xlsx")
}
