% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_display_demo.R
\name{launch_display_demo}
\alias{launch_display_demo}
\title{launch_display_demo}
\usage{
launch_display_demo()
}
\value{
launch_display_demo app
}
\description{
demo of the display module (\code{mod_display_ui()}/\code{mod_display_server()})
}
