% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_upload_demo.R
\name{launch_upload_demo}
\alias{launch_upload_demo}
\title{launch_upload_demo}
\usage{
launch_upload_demo()
}
\value{
app
}
\description{
demo of the upload module (\code{mod_upload_ui()}/\code{mod_upload_server()})
}
