% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_flat_file.R
\name{load_flat_file}
\alias{load_flat_file}
\title{Load flat data files}
\usage{
load_flat_file(path)
}
\arguments{
\item{path}{path to data file (with extension)}
}
\value{
return_data
}
\description{
Load flat data files
}
\examples{
# from local
load_flat_file(path = "inst/extdata/csv/2015-baseballdatabank/core/AllstarFull.csv")

}
