% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_compare.R
\name{mod_compare_ui}
\alias{mod_compare_ui}
\title{mod_compare_ui}
\usage{
mod_compare_ui(id, dev = FALSE)
}
\arguments{
\item{id}{module id}

\item{dev}{show developer reactive-value outputs (default FALSE)}
}
\description{
compare UI module
}
