% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_display.R
\name{mod_display_ui}
\alias{mod_display_ui}
\title{mod_display_ui}
\usage{
mod_display_ui(id)
}
\arguments{
\item{id}{module id}
}
\value{
display data UI
}
\description{
UI module that previews the uploaded base and compare data
}
