% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_select.R
\name{mod_select_server}
\alias{mod_select_server}
\title{mod_select_server}
\usage{
mod_select_server(id, data_upload, dev = FALSE)
}
\arguments{
\item{id}{module id}

\item{dev}{show developer reactive-value outputs (default FALSE)}
}
\description{
select server module
}
