% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload.R
\name{mod_upload_ui}
\alias{mod_upload_ui}
\title{mod_upload_ui()}
\usage{
mod_upload_ui(id, dev = FALSE)
}
\arguments{
\item{id}{module id}

\item{dev}{show developer reactive-value outputs (default FALSE)}
}
\value{
A \code{tagList} of the upload module's UI (file inputs and preview
tables for \code{base} and \code{compare})
}
\description{
UI module for upload
}
