% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react_themes.R
\name{new_react_theme}
\alias{new_react_theme}
\title{reactable theme (new data)}
\usage{
new_react_theme(dark = FALSE)
}
\arguments{
\item{dark}{use dark-mode colors (default FALSE)}
}
\description{
theme for new data in reactable table
}
