% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_columns.R
\name{print.dfdiffs_column_diff}
\alias{print.dfdiffs_column_diff}
\title{Print a dfdiffs_column_diff object}
\usage{
\method{print}{dfdiffs_column_diff}(x, ...)
}
\arguments{
\item{x}{a \code{dfdiffs_column_diff} object}

\item{...}{unused}
}
\value{
\code{x}, invisibly
}
\description{
Print a dfdiffs_column_diff object
}
