% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_summary.R
\name{print.dfdiffs_summary}
\alias{print.dfdiffs_summary}
\title{Print a dfdiffs_summary object}
\usage{
\method{print}{dfdiffs_summary}(x, ...)
}
\arguments{
\item{x}{a \code{dfdiffs_summary} object}

\item{...}{unused}
}
\value{
\code{x}, invisibly
}
\description{
Print a dfdiffs_summary object
}
