% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_compare_df.R
\name{proc_compare_df}
\alias{proc_compare_df}
\title{Perform comparison on two datasets}
\usage{
proc_compare_df(data_x, data_y, by, by_col = NULL, cols = NULL)
}
\arguments{
\item{data_x}{'current' or 'new' dataset}

\item{data_y}{'previous' or 'old' dataset}

\item{by}{supplied join columns}

\item{by_col}{unique identifier}

\item{cols}{columns to compare}
}
\value{
compdf_list list of tables
}
\description{
Perform comparison on two datasets
}
\examples{
# using dfdiffs::diff_modified_data
diff_modified_data <- dfdiffs::diff_modified_data
diff_current_modified <- diff_modified_data$diff_current_modified
diff_previous_modified <- diff_modified_data$diff_previous_modified
proc_compare_df(data_x = diff_current_modified,
                data_y = diff_previous_modified,
                by = c("subject_id", "record"),
                cols = c("text_value_a", "text_value_b"))
}
