% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Roster2022}
\alias{Roster2022}
\title{Roster2022 (synthetic site roster, year two)}
\format{
A data frame with 237 rows and 11 variables (see \code{\link{Roster2021}})
}
\usage{
Roster2022
}
\description{
\code{Roster2021} plus 40 newly enrolled subjects, minus 3 who withdrew.
See \code{\link{Roster2021}} for column details.
}
\keyword{datasets}
