% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadDataUI.R
\name{uploadDataUI}
\alias{uploadDataUI}
\title{data upload (UI)}
\usage{
uploadDataUI(id)
}
\arguments{
\item{id}{module id}
}
\description{
data upload (UI)
}
\examples{
# build ui for data upload
uploadDataUI("input_id")
}
