% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react_themes.R
\docType{data}
\name{base_react_theme}
\alias{base_react_theme}
\title{reactable theme (base data)}
\format{
An object of class \code{reactableTheme} of length 9.
}
\usage{
base_react_theme
}
\description{
theme for base data in reactable table
}
\keyword{datasets}
