% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react_themes.R
\docType{data}
\name{changed_react_theme}
\alias{changed_react_theme}
\title{reactable theme (changed data)}
\format{
An object of class \code{reactableTheme} of length 9.
}
\usage{
changed_react_theme
}
\description{
theme for changed data in reactable table
}
\keyword{datasets}
