% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ChangedData}
\alias{ChangedData}
\title{Changed/Modified Current Data (simulated data for checking modified data)}
\format{
A data frame with 5 rows and 7 variables:
\describe{
\item{subject_id}{A character ID variable ("A", "B", "C")}
\item{record}{A numeric record ID (1-5)}
\item{text_value_a}{a variable containing random text}
\item{text_value_b}{a variable containing random text}
\item{created_date}{a beginning date (from 2021-07-16  to 2021-08-24)}
\item{updated_date}{a middle date (from 2021-10-03 to 2021-11-27)}
\item{entered_date}{a last date (from 2021-11-11 to 2021-11-30)}
}
}
\usage{
ChangedData
}
\description{
A dataset containing five rows and seven columns. The variables are as
follows:
}
\keyword{datasets}
