% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react_themes.R
\docType{data}
\name{comp_react_theme}
\alias{comp_react_theme}
\title{reactable theme (compare data)}
\format{
An object of class \code{reactableTheme} of length 9.
}
\usage{
comp_react_theme
}
\description{
theme for compare data in reactable table
}
\keyword{datasets}
