% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_data.R
\name{compare_data}
\alias{compare_data}
\title{Compare data (create new, deleted, and changed data)}
\usage{
compare_data(compare, base, by = NULL, by_col = NULL, cols = NULL)
}
\arguments{
\item{compare}{comparison data table}

\item{base}{base data table}

\item{by}{join column}

\item{by_col}{new join column name}

\item{cols}{columns to compare}
}
\value{
list of comparison tables
}
\description{
Compare data (create new, deleted, and changed data)
}
\examples{
# not run
r21 <- dfdiffs::Roster2021
r22 <- dfdiffs::Roster2022
compare_data(compare = r22, base = r21,
    by = "subject_id",
    cols = c("first_name", "last_name", "full_name"))
}
