% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_comparison_report.R
\name{create_comparison_report}
\alias{create_comparison_report}
\title{Create comparison report}
\usage{
create_comparison_report(
  compare,
  base,
  by = NULL,
  by_col = NULL,
  cols = NULL,
  file
)
}
\arguments{
\item{compare}{compare df}

\item{base}{base df}

\item{by}{keys or joining bs4Dash::column}

\item{by_col}{bs4Dash::column name for joining bs4Dash::column}

\item{cols}{columns to compare}

\item{file}{export file (with path)}
}
\value{
xlsx workbook
}
\description{
Create comparison report
}
