% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_modified_data.R
\name{create_modified_data}
\alias{create_modified_data}
\title{Create modified data}
\usage{
create_modified_data(compare, base, by = NULL, by_col = NULL, cols = NULL)
}
\arguments{
\item{compare}{A 'current' or 'new' dataset (tibble or data.frame)}

\item{base}{A 'previous' or 'old' dataset (tibble or data.frame)}

\item{by}{A join column between the two datasets, or any combination of columns that constitute a unique row.}

\item{by_col}{A new name for the joining column.}

\item{cols}{Columns to be compared.}
}
\value{
a list with \code{diffs} (long format: \verb{Variable name}, key column(s),
\verb{Current Value}, \verb{Previous Value}) and \code{diffs_byvar} (count of
differing values per variable)
}
\description{
Create modified data
}
\examples{
# with local data
CurrentData <- dfdiffs::ChangedData
PreviousData <- dfdiffs::InitialData
create_modified_data(
           compare = CurrentData,
           base = PreviousData,
           by = c("subject_id", "record"),
           cols = c("text_value_a", "text_value_b", "updated_date"))
create_modified_data(
           compare = CurrentData,
           base = PreviousData,
           by = c("subject_id", "record"))
}
