% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_column.R
\name{create_new_column}
\alias{create_new_column}
\title{Create new (joining) column}
\usage{
create_new_column(data, cols, new_name, sep)
}
\arguments{
\item{data}{a tibble or data.frame}

\item{cols}{cols to create new column from (they will be pasted together with "-")}

\item{new_name}{new column name}
}
\value{
new_col_data data with new column
}
\description{
Create new (joining) column
}
\examples{
library(dplyr)
library(tidyr)
CompleteData <- dfdiffs::CompleteData
IncompleteData <- dfdiffs::IncompleteData
CompleteDataJoin <- create_new_column(data = CompleteData,
                                       cols = c("subject", "record"),
                                       new_name = "join_var")
IncompleteDataJoin <- create_new_column(data = IncompleteData,
                                       cols = c("subject", "record"),
                                       new_name = "join_var")
}
