% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react_themes.R
\docType{data}
\name{deleted_react_theme}
\alias{deleted_react_theme}
\title{reactable theme (deleted data)}
\format{
An object of class \code{reactableTheme} of length 9.
}
\usage{
deleted_react_theme
}
\description{
theme for deleted data in reactable table
}
\keyword{datasets}
