% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfdiffs_fresh_theme.R
\name{dfdiffs_fresh_theme}
\alias{dfdiffs_fresh_theme}
\title{custom theme (fresh <> bs4Dash)}
\usage{
dfdiffs_fresh_theme()
}
\value{
theme shiny app
}
\description{
this is the fresh theme with custom colors.
}
