% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app_dev.R
\name{launch_app_dev}
\alias{launch_app_dev}
\title{launch_app_dev}
\usage{
launch_app_dev()
}
\description{
dev variant of \code{launch_app()} with the modules' debug
reactive-value outputs turned on (\code{dev = TRUE})
}
\examples{
\dontrun{
launch_app_dev()
}
}
