% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_select_demo.R
\name{launch_select_demo}
\alias{launch_select_demo}
\title{launch_select_demo}
\usage{
launch_select_demo()
}
\description{
demo of the select module (\code{mod_select_ui()}/\code{mod_select_server()})
}
\examples{
\dontrun{
launch_select_demo()
}
}
