% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_data.R
\name{load_flat_file}
\alias{load_flat_file}
\title{Load flat data files}
\usage{
load_flat_file(path)
}
\arguments{
\item{path}{path to data file (with extension)}
}
\value{
return_data as loaded flat file
}
\description{
Load flat data files
}
