% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_compare.R
\name{mod_compare_server}
\alias{mod_compare_server}
\title{mod_compare_server}
\usage{
mod_compare_server(id, data_selected, dev = FALSE)
}
\arguments{
\item{id}{module id}

\item{dev}{show developer reactive-value outputs (default FALSE)}
}
\description{
compare server module
}
