% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_select.R
\name{mod_select_ui}
\alias{mod_select_ui}
\title{mod_select_ui}
\usage{
mod_select_ui(id, dev = FALSE)
}
\arguments{
\item{id}{module id}

\item{dev}{show developer reactive-value outputs (default FALSE)}
}
\value{
A \code{tagList} of the select module's UI (column selection and
join-column accordions)
}
\description{
select UI module
}
