% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react_themes.R
\docType{data}
\name{new_react_theme}
\alias{new_react_theme}
\title{reactable theme (new data)}
\format{
An object of class \code{reactableTheme} of length 9.
}
\usage{
new_react_theme
}
\description{
theme for new data in reactable table
}
\keyword{datasets}
