% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NewData}
\alias{NewData}
\title{NewData (difference between 'time-point 1' & 'time-point 2' data)}
\format{
A data frame with 3 rows and 7 variables:
\describe{
\item{subject}{A character ID variable ("A", "B", "D")}
\item{record}{A numeric record ID (2, 4, 5)}
\item{start_date}{a beginning date (from 2022-04-02 to 2022-04-04)}
\item{mid_date}{a middle date (from 2022-04-13 to 2022-04-15)}
\item{end_date}{a last date (from 2022-04-20 to 2022-04-22)}
\item{text_var}{a variable containing random text}
\item{factor_var}{a categorical variable formatted as a factor}
}
}
\usage{
NewData
}
\description{
A dataset containing three rows and seven columns. The variables are as
follows:
}
\keyword{datasets}
