% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{T1Data}
\alias{T1Data}
\title{T1Data (simulated 'time-point 1' data for checking 'new data')}
\format{
A data frame with 6 rows and 7 variables:
\describe{
\item{subject}{A character ID variable ("A", "B", "C", "D")}
\item{record}{A numeric record ID (1-6)}
\item{start_date}{a beginning date (from 2022-01-25 to 2022-01-30)}
\item{mid_date}{a middle date (from 2022-03-15 to 2022-03-20)}
\item{end_date}{a last date (from 2022-03-25 to 2022-03-31)}
\item{text_var}{a variable containing random text}
\item{factor_var}{a categorical variable formatted as a factor}
}
}
\usage{
T1Data
}
\description{
A dataset containing six rows and seven columns. The variables are as
follows:
}
\keyword{datasets}
