% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_data.R
\name{upload_data}
\alias{upload_data}
\title{Upload data files}
\usage{
upload_data(path, sheet = NULL)
}
\arguments{
\item{path}{path to data file (with extension)}

\item{sheet}{excel sheet (if excel file)}
}
\value{
uploaded data file
}
\description{
Upload data files
}
